refactor(web): give the voice placeholder and section their own files
One component per file, per the web style guide: the placeholder carries its
own responsibility rather than being a trivial helper private to the page.
`VoiceSection` moves alongside it because both render it, and importing it
back from the page would close an import cycle. `CardSkeleton` stays private
to the placeholder as the sanctioned exception.
